move delete tests to action_delete_test.go
